pmid2abstract

pmid2abstract

Usage

pmid2abstract(pmid)

Arguments

pmid

Examples

pmid2abstract(33317393)
#> [1] "Introduction: Working memory (WM) and its blood-oxygen-level-dependent-related parametric modulation under load decrease with age. Functional connectivity (FC) generally increases with WM load; however, how aging impacts connectivity and whether this is load-dependent, region-dependent, or associated with cognitive performance is unclear. Methods: This study examines these questions in 170 healthy adults (meanage = 52.99 ± 19.18) who completed functional magnetic resonance imaging scanning during an n-back task (0-, 2-, 3-, and 4-back). The FC was estimated by utilizing a modified generalized psychophysiological interaction approach with seeds from fronto-parietal (FP) and default mode (DM) regions that modulated to n-back difficulty. The FC analyses focused on both connectivity during WM engagement (task vs. control) and connectivity in response to increased WM load (linear slope across conditions). Each analysis utilized within- and between-region FC, predicted by age (linear or quadratic), and its associations with in- and out-of-scanner task performance. Results: Engaging in WM either generally (task vs. control) or as a function of difficulty strengthened integration within- and between-FP and DM regions. Notably, these task-sensitive functional connections were robust to the effects of age. Stronger negative FC between FP and DM regions was also associated with better WM performance in an age-dependent manner, occurring selectively in middle-aged and older adults. Discussion: These results suggest that FC is critical for engaging in cognitively demanding tasks, and its lack of sensitivity to healthy aging may provide a means to maintain cognition across the adult lifespan. Thus, this study highlights the contribution of maintenance in brain function to support working memory processing with aging."